third_party/WebKit/LayoutTests/webaudio/audionode-expected.txt

Basic tests for AudioNode API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Source AudioNode has no inputs.
PASS Source AudioNode has one output.
PASS Destination AudioNode has one input.
PASS Destination AudioNode has no outputs.
PASS connect() exception thrown for illegal destination AudioNode.
PASS connect() exception thrown for illegal output index.
PASS connect() exception thrown for illegal input index.
PASS audioNode.connect(context.destination) succeeded.
PASS exception thrown when connecting to other context's node.
PASS context3 = new AudioContext(1, 44100, 44100) did not throw exception.
PASS context3 is not an OfflineAudioContext
PASS AudioNode is an EventTarget
PASS successfullyParsed is true

TEST COMPLETE
